fix: add bytecode, move query
