+++ packages/packages.json
@@ -1,15 +1,15 @@
 {
     "dev": {
-        "skill/valory/proposal_collector_abci/0.1.0": "bafybeiaisb724vozayjyquzmrmgnsmcfassxfnoa7g3hxu6chcblaxpoqy",
+        "skill/valory/proposal_collector_abci/0.1.0": "bafybeigzunepbfo7srdsb572gb5scnq5klxrzmkgwekq35sisoo4xhxcou",
         "connection/valory/openai/0.1.0": "bafybeietimatztynqebwicxabplia6fctov4vswzbdrin5y32brbjewo2y",
         "contract/valory/delegate/0.1.0": "bafybeiburzphuq5b3bsmhe4s3mvxuej4uhnexlsuxwgmujqaiisut37kpm",
         "protocol/valory/llm/1.0.0": "bafybeicpd5hxe5rlgp75kaqotzpy3jkdyl5pbdcutwqdzlqfknnmkgadsy",
-        "skill/valory/proposal_voter_abci/0.1.0": "bafybeihlantthkkzcwpdf7bfhwdy6yk2gqieldsj555tk3pbua3hm7ecu4",
-        "skill/valory/governatooorr_abci/0.1.0": "bafybeihjpydaifovabrsgeiqja5cz77clcwoagwditpsaa5snofqirugqe",
-        "agent/valory/governatooorr/0.1.0": "bafybeibl3elvidx7noatfq7h5uuwn6mmfzls7ett3dvsxdsage6qyzvv54",
-        "service/valory/governatooorr/0.1.0": "bafybeic2opeziiiwr2bbyigmkxfl6w76me3uwikvq3essqwrjl73ebskbi",
-        "service/valory/governatooorr_local/0.1.0": "bafybeigui6udxgmu2oo65f27c6jl36bwh5e5gtipvxabsofxkbe2rnxh4e",
-        "contract/valory/sign_message_lib/0.1.0": "bafybeibd2do7fa4mem53a7jyo2avrnsrmq3r4spumcexrtxh7fxpnkjuoi"
+        "skill/valory/proposal_voter_abci/0.1.0": "bafybeia6usdfg5asih23v7iac7mxflr3uybr3sif46gwa2gcseffqowasm",
+        "skill/valory/governatooorr_abci/0.1.0": "bafybeiftclp6g32ihgoavrw7rx4vrd4mjhx5er5vpu4bcoagblicrszomm",
+        "agent/valory/governatooorr/0.1.0": "bafybeibnnnwlnmhks5qeaqzvtwov4fteywqatfri27gzonsfb7bwry7cgi",
+        "service/valory/governatooorr/0.1.0": "bafybeiecj2ulb7cfzmvnrgofjvepowq4qgjfw3fpbglslypymbkmgonjhe",
+        "service/valory/governatooorr_local/0.1.0": "bafybeihln73rox62taznprjmqvjkxqu2v2bykjc4srtu3dlenbfcxuufse",
+        "contract/valory/sign_message_lib/0.1.0": "bafybeibblmummqukegwkgwv5gsgrh4keaikmvnsk4dodhcuezxw7hi67yi"
     },
     "third_party": {
         "protocol/open_aea/signing/1.0.0": "bafybeibqlfmikg5hk4phzak6gqzhpkt6akckx7xppbp53mvwt6r73h7tk4",
