+++ packages/valory/agents/governatooorr/aea-config.yaml
@@ -24,6 +24,7 @@ contracts:
 - valory/delegate:0.1.0:bafybeiburzphuq5b3bsmhe4s3mvxuej4uhnexlsuxwgmujqaiisut37kpm
 - valory/gnosis_safe:0.1.0:bafybeic3cfap5mbdfb67iejg6zaobtpgvxoltckdgbeovwc7gpzvjz2md4
 - valory/gnosis_safe_proxy_factory:0.1.0:bafybeif3gj5a5n6humzzsc3hacmfvfdqltzo7ibc7vvtisv3ppmbzxbkui
+- valory/sign_message_lib:0.1.0:bafybeibblmummqukegwkgwv5gsgrh4keaikmvnsk4dodhcuezxw7hi67yi
 protocols:
 - open_aea/signing:1.0.0:bafybeibqlfmikg5hk4phzak6gqzhpkt6akckx7xppbp53mvwt6r73h7tk4
 - valory/abci:0.1.0:bafybeig3dj5jhsowlvg3t73kgobf6xn4nka7rkttakdb2gwsg5bp7rt7q4
@@ -37,9 +38,9 @@ protocols:
 skills:
 - valory/abstract_abci:0.1.0:bafybeiet34ju7typtkqyfn572sacef5krrsb6d4vpfnviqi6kzunje3xny
 - valory/abstract_round_abci:0.1.0:bafybeibdljcw2wchyderkonobpdpyb36yioagqcaoqldydw5bi6vvempma
-- valory/governatooorr_abci:0.1.0:bafybeihjpydaifovabrsgeiqja5cz77clcwoagwditpsaa5snofqirugqe
-- valory/proposal_collector_abci:0.1.0:bafybeiaisb724vozayjyquzmrmgnsmcfassxfnoa7g3hxu6chcblaxpoqy
-- valory/proposal_voter_abci:0.1.0:bafybeihlantthkkzcwpdf7bfhwdy6yk2gqieldsj555tk3pbua3hm7ecu4
+- valory/governatooorr_abci:0.1.0:bafybeiftclp6g32ihgoavrw7rx4vrd4mjhx5er5vpu4bcoagblicrszomm
+- valory/proposal_collector_abci:0.1.0:bafybeigzunepbfo7srdsb572gb5scnq5klxrzmkgwekq35sisoo4xhxcou
+- valory/proposal_voter_abci:0.1.0:bafybeia6usdfg5asih23v7iac7mxflr3uybr3sif46gwa2gcseffqowasm
 - valory/registration_abci:0.1.0:bafybeidqcszsitlmsofegmbprny22hqlent2o6luxng5a46gplytexrtbq
 - valory/reset_pause_abci:0.1.0:bafybeidprdrxzbloz5imufkbpvetvkhywdvq3fopbhdzztfnud4onsdsce
 - valory/termination_abci:0.1.0:bafybeif33wyu6w7yrrhyacpt627i5g6nec2kiepsm5xx2b44zq657dch3a
