+++ packages/valory/services/governatooorr/service.yaml
@@ -8,7 +8,7 @@ license: Apache-2.0
 fingerprint:
   README.md: bafybeid6vqhnovhpsrjwws2wcb67i7fwr4dq54rpowqgrf3njt2anp5nc4
 fingerprint_ignore_patterns: []
-agent: valory/governatooorr:0.1.0:bafybeibl3elvidx7noatfq7h5uuwn6mmfzls7ett3dvsxdsage6qyzvv54
+agent: valory/governatooorr:0.1.0:bafybeibnnnwlnmhks5qeaqzvtwov4fteywqatfri27gzonsfb7bwry7cgi
 number_of_agents: 4
 deployment:
   agent:
