+++ packages/valory/skills/proposal_collector_abci/skill.yaml
@@ -15,7 +15,7 @@ fingerprint:
   models.py: bafybeigrf4yztnaref5jgp2wpqlg2osrib42p4axymcstfrli6wga3n2kq
   payloads.py: bafybeigpm6xtiddrc5fqyivjttpqkwn5fqnag4xo3c3fz5awne3txr3hg4
   rounds.py: bafybeib6zul7we7wvhpa43smk2wsle7dj7dvvxlcd73hkb3zxwvu6uhedi
-  snapshot.py: bafybeig6c3yg2py6i233itwfzcekquelj34dopwt3moltl6vekpbmy47ya
+  snapshot.py: bafybeiccroyc43n3nd3v5qjlxaenef5sqfakq6nacwgdn3n6khi4cdjs54
   tally.py: bafybeih4w2gk6vzthjbw3q2hhckquxgcs7y4kvj3uezclugb6qvo3s35fq
   tests/__init__.py: bafybeic43gkt67rz4d2latl5qvjxymqjvim4mcj6aklxa6bl73yobtopra
   tests/test_behaviours.py: bafybeibthjmkcgyyg6wdratk2f6rrsfj65iztr5kbzx4xs6thdpuagz3yq
